refactor: split handle message api
