+++ packages/client-direct/src/index.ts
@@ -26,6 +26,7 @@ import { createApiRouter } from './api.ts';
 import * as fs from 'fs';
 import * as path from 'path';
 import OpenAI from 'openai';
+import {getUserMessage, handleUserMessage} from "./process.ts";
 
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {
@@ -207,527 +208,21 @@ export class DirectClient {
             '/:agentId/message',
             upload.single('file'),
             async (req: express.Request, res: express.Response) => {
-                const messageStart = Date.now();
                 const agentId = req.params.agentId;
-                const roomId = stringToUuid(
-                    req.body.roomId ?? 'default-room-' + agentId,
-                );
-                const accessToken = req.body?.accessToken;
-                if (process.env?.ENABLE_CHAT_AUTH == 'true' && !accessToken) {
-                    res.status(401).send('No accessToken provided');
-                    return;
-                }
-                const userId = stringToUuid(req.body.userId ?? 'user');
-
                 let runtime = this.agents.get(agentId);
-
-                // if runtime is null, look for runtime with the same name
-                if (!runtime) {
-                    runtime = Array.from(this.agents.values()).find(
-                        (a) =>
-                            a.character.name.toLowerCase() ===
-                            agentId.toLowerCase(),
-                    );
-                }
-
                 if (!runtime) {
                     res.status(404).send('Agent not found');
                     return;
                 }
-
-                await runtime.ensureConnection(
-                    userId,
-                    roomId,
-                    req.body.userName,
-                    req.body.name,
-                    'direct',
-                );
-
-                const text = req.body.text;
-
-                const messageId = stringToUuid(Date.now().toString());
-
-                const attachments: Media[] = [];
-                if (req.file) {
-                    const filePath = path.join(
-                        process.cwd(),
-                        'data',
-                        'uploads',
-                        req.file.filename,
-                    );
-                    attachments.push({
-                        id: Date.now().toString(),
-                        url: filePath,
-                        title: req.file.originalname,
-                        source: 'direct',
-                        description: `Uploaded file: ${req.file.originalname}`,
-                        text: '',
-                        contentType: req.file.mimetype,
-                    });
+                if (process.env?.ENABLE_CHAT_AUTH == 'true' && !req.body.accessToken) {
+                    res.status(401).send('No accessToken provided');
+                    return;
                 }
-
-                const content: Content = {
-                    text,
-                    attachments,
-                    source: 'direct',
-                    accessToken: accessToken,
-                    inReplyTo: undefined,
-                };
-
-                const userMessage = {
-                    content,
-                    userId,
-                    roomId,
-                    agentId: runtime.agentId,
-                    createdAt: Date.now(),
-                };
-
-                const memory: Memory = {
-                    id: stringToUuid(messageId + '-' + userId),
-                    ...userMessage,
-                    agentId: runtime.agentId,
-                    userId,
-                    roomId,
-                    content,
-                    createdAt: Date.now(),
-                };
-
-                memory.embedding = await getEmbeddingZeroVector();
-                await runtime.messageManager.createMemory(memory);
-                let state = await runtime.composeState(userMessage, {
-                    agentName: runtime.character.name,
-                    conversationLength: 10,
-                });
-
-                const responseMessages = [];
-                // query tasks
-                let task_record = {
-                    roomId,
-                    agentId,
-                    userId,
-                    taskId: 0,
-                    taskDefinition: text,
-                    pastActions: [],
-                };
-                if (
-                    'queryLatestTask' in runtime.databaseAdapter &&
-                    typeof runtime.databaseAdapter.queryLatestTask ===
-                        'function'
-                ) {
-                    const query = {
-                        roomId,
-                        agentId,
-                        userId,
-                    };
-                    const result =
-                        await runtime.databaseAdapter.queryLatestTask(
-                            'tasks',
-                            query,
-                        );
-                    console.log('result', JSON.stringify(result));
-                    const lastestTask = result?.[0];
-                    console.log('lastestTask', JSON.stringify(lastestTask));
-                    if (
-                        lastestTask?.pastActions?.length > 0 &&
-                        lastestTask.pastActions[
-                            lastestTask.pastActions.length - 1
-                        ]?.action === 'WRAP_UP'
-                    ) {
-                        task_record.taskId = lastestTask.taskId + 1;
-                    } else if (lastestTask) {
-                        task_record = lastestTask;
-                    }
-                }
-                console.log('task_record', task_record);
-
-                // get recent messages
-                const recentMessages = await runtime.messageManager.getMemories(
-                    {
-                        roomId,
-                        count: 5,
-                        unique: false,
-                    },
-                );
-
-                // convert into the request format of agent-router
-                console.log('recentMessages', recentMessages);
-                const chatHistory = recentMessages
-                    .slice()
-                    .reverse()
-                    .map((msg, idx) => {
-                        const role =
-                            msg.userId === runtime.agentId
-                                ? 'assistant'
-                                : 'user';
-                        return {
-                            role,
-                            content: msg.content.text || '',
-                            idx: idx,
-                        };
-                    });
-                console.log('chatHistory', chatHistory);
-                for (let stepCnt = 0; stepCnt < 5; stepCnt++) {
-                    let shouldReturn = false;
-
-                    // call agent router to get response
-                    const body = {
-                        chat_history: chatHistory,
-                        task_definition: task_record.taskDefinition,
-                        past_steps: task_record?.pastActions || [],
-                    };
-                    console.log('body', body);
-                    const agentRouterResponse = await fetch(
-                        runtime.getSetting('AGENT_ROUTER_URL') + '/plan',
-                        {
-                            method: 'POST',
-                            headers: {
-                                'Content-Type': 'application/json',
-                            },
-                            body: JSON.stringify(body),
-                        },
-                    ).then((res) => res.json());
-
-                    console.log(
-                        `${messageId} message query elapsed: ${Date.now() - messageStart}ms, response: ${JSON.stringify(agentRouterResponse)}`,
-                    );
-
-                    // save response to memory
-                    const aiResponseMemory: Memory = {
-                        id: stringToUuid(Date.now().toString()),
-                        ...userMessage,
-                        userId: runtime.agentId,
-                        agentId: runtime.agentId,
-                        content: agentRouterResponse,
-                        embedding: getEmbeddingZeroVector(),
-                        createdAt: Date.now(),
-                    };
-
-                    let actionResponseMessage = null as Content | null;
-                    if (agentRouterResponse.action === 'WRAP_UP') {
-                        actionResponseMessage = {
-                            text: agentRouterResponse.parameters.message,
-                            action: agentRouterResponse.action,
-                        };
-                        shouldReturn = true;
-                    } else {
-                        state = await runtime.composeState(userMessage, {
-                            agentName: runtime.character.name,
-                            actionParameters: agentRouterResponse['parameters'],
-                        });
-                        const actionsProcessResult =
-                            await runtime.processActions(
-                                memory,
-                                [aiResponseMemory],
-                                state,
-                                async (newMessages) => {
-                                    actionResponseMessage = newMessages;
-                                    return [memory];
-                                },
-                            );
-
-                        // return to user if any action failed
-                        console.log(
-                            'actionsProcessResult',
-                            actionsProcessResult,
-                        );
-                        for (const processResult of actionsProcessResult) {
-                            if (processResult === false) {
-                                shouldReturn = true;
-                                break;
-                            }
-                        }
-                    }
-                    console.log('actionResponseMessage', actionResponseMessage);
-                    task_record.pastActions.push({
-                        action: agentRouterResponse.action,
-                        detail: agentRouterResponse.explanation,
-                        result:
-                            actionResponseMessage?.result ||
-                            actionResponseMessage?.text,
-                    });
-
-                    if (
-                        agentRouterResponse.action === 'WRAP_UP' ||
-                        shouldReturn === true
-                    ) {
-                        responseMessages.push(actionResponseMessage);
-                        const resMemory: Memory = {
-                            id: stringToUuid(Date.now().toString()),
-                            roomId: userMessage.roomId,
-                            userId: runtime.agentId,
-                            agentId: runtime.agentId,
-                            content: actionResponseMessage,
-                            embedding: getEmbeddingZeroVector(),
-                            createdAt: Date.now(),
-                        };
-                        console.log('resMemory', resMemory);
-                        state = await runtime.updateRecentMessageState(state);
-                        await runtime.messageManager.createMemory(
-                            resMemory,
-                            true,
-                        );
-                    }
-
-                    if (shouldReturn) {
-                        break;
-                    }
-                }
-
-                if (
-                    'upsert' in runtime.databaseAdapter &&
-                    typeof runtime.databaseAdapter.upsert === 'function'
-                ) {
-                    runtime.databaseAdapter.upsert('tasks', task_record);
-                }
+                const userMessage = await getUserMessage(runtime, req);
+                const responseMessages = await handleUserMessage(runtime, userMessage);
                 res.json(responseMessages);
             },
         );
-
-        // this.app.post(
-        //     "/agents/:agentIdOrName/hyperfi/v1",
-        //     async (req: express.Request, res: express.Response) => {
-        //         // get runtime
-        //         const agentId = req.params.agentIdOrName;
-        //         let runtime = this.agents.get(agentId);
-        //         // if runtime is null, look for runtime with the same name
-        //         if (!runtime) {
-        //             runtime = Array.from(this.agents.values()).find(
-        //                 (a) =>
-        //                     a.character.name.toLowerCase() ===
-        //                     agentId.toLowerCase()
-        //             );
-        //         }
-        //         if (!runtime) {
-        //             res.status(404).send("Agent not found");
-        //             return;
-        //         }
-        //
-        //         // can we be in more than one hyperfi world at once
-        //         // but you may want the same context is multiple worlds
-        //         // this is more like an instanceId
-        //         const roomId = stringToUuid(req.body.roomId ?? "hyperfi");
-        //
-        //         const body = req.body;
-        //
-        //         // hyperfi specific parameters
-        //         let nearby = [];
-        //         let availableEmotes = [];
-        //
-        //         if (body.nearby) {
-        //             nearby = body.nearby;
-        //         }
-        //         if (body.messages) {
-        //             // loop on the messages and record the memories
-        //             // might want to do this in parallel
-        //             for (const msg of body.messages) {
-        //                 const parts = msg.split(/:\s*/);
-        //                 const mUserId = stringToUuid(parts[0]);
-        //                 await runtime.ensureConnection(
-        //                     mUserId,
-        //                     roomId, // where
-        //                     parts[0], // username
-        //                     parts[0], // userScreeName?
-        //                     "hyperfi"
-        //                 );
-        //                 const content: Content = {
-        //                     text: parts[1] || "",
-        //                     attachments: [],
-        //                     source: "hyperfi",
-        //                     inReplyTo: undefined,
-        //                 };
-        //                 const memory: Memory = {
-        //                     id: stringToUuid(msg),
-        //                     agentId: runtime.agentId,
-        //                     userId: mUserId,
-        //                     roomId,
-        //                     content,
-        //                 };
-        //                 await runtime.messageManager.createMemory(memory);
-        //             }
-        //         }
-        //         if (body.availableEmotes) {
-        //             availableEmotes = body.availableEmotes;
-        //         }
-        //
-        //         const content: Content = {
-        //             // we need to compose who's near and what emotes are available
-        //             text: JSON.stringify(req.body),
-        //             attachments: [],
-        //             source: "hyperfi",
-        //             inReplyTo: undefined,
-        //         };
-        //
-        //         const userId = stringToUuid("hyperfi");
-        //         const userMessage = {
-        //             content,
-        //             userId,
-        //             roomId,
-        //             agentId: runtime.agentId,
-        //         };
-        //
-        //         const state = await runtime.composeState(userMessage, {
-        //             agentName: runtime.character.name,
-        //         });
-        //
-        //         let template = hyperfiHandlerTemplate;
-        //         template = template.replace(
-        //             "{{emotes}}",
-        //             availableEmotes.join("|")
-        //         );
-        //         template = template.replace("{{nearby}}", nearby.join("|"));
-        //         const context = composeContext({
-        //             state,
-        //             template,
-        //         });
-        //
-        //         function createHyperfiOutSchema(
-        //             nearby: string[],
-        //             availableEmotes: string[]
-        //         ) {
-        //             const lookAtSchema =
-        //                 nearby.length > 1
-        //                     ? z
-        //                           .union(
-        //                               nearby.map((item) => z.literal(item)) as [
-        //                                   z.ZodLiteral<string>,
-        //                                   z.ZodLiteral<string>,
-        //                                   ...z.ZodLiteral<string>[]
-        //                               ]
-        //                           )
-        //                           .nullable()
-        //                     : nearby.length === 1
-        //                     ? z.literal(nearby[0]).nullable()
-        //                     : z.null(); // Fallback for empty array
-        //
-        //             const emoteSchema =
-        //                 availableEmotes.length > 1
-        //                     ? z
-        //                           .union(
-        //                               availableEmotes.map((item) =>
-        //                                   z.literal(item)
-        //                               ) as [
-        //                                   z.ZodLiteral<string>,
-        //                                   z.ZodLiteral<string>,
-        //                                   ...z.ZodLiteral<string>[]
-        //                               ]
-        //                           )
-        //                           .nullable()
-        //                     : availableEmotes.length === 1
-        //                     ? z.literal(availableEmotes[0]).nullable()
-        //                     : z.null(); // Fallback for empty array
-        //
-        //             return z.object({
-        //                 lookAt: lookAtSchema,
-        //                 emote: emoteSchema,
-        //                 say: z.string().nullable(),
-        //                 actions: z.array(z.string()).nullable(),
-        //             });
-        //         }
-        //
-        //         // Define the schema for the expected output
-        //         const hyperfiOutSchema = createHyperfiOutSchema(
-        //             nearby,
-        //             availableEmotes
-        //         );
-        //
-        //         // Call LLM
-        //         const response = await generateObject({
-        //             runtime,
-        //             context,
-        //             modelClass: ModelClass.SMALL, // 1s processing time on openai small
-        //             schema: hyperfiOutSchema,
-        //         });
-        //
-        //         if (!response) {
-        //             res.status(500).send(
-        //                 "No response from generateMessageResponse"
-        //             );
-        //             return;
-        //         }
-        //
-        //         let hfOut;
-        //         try {
-        //             hfOut = hyperfiOutSchema.parse(response.object);
-        //         } catch {
-        //             elizaLogger.error(
-        //                 "cant serialize response",
-        //                 response.object
-        //             );
-        //             res.status(500).send("Error in LLM response, try again");
-        //             return;
-        //         }
-        //
-        //         // do this in the background
-        //         new Promise((resolve) => {
-        //             const contentObj: Content = {
-        //                 text: hfOut.say,
-        //             };
-        //
-        //             if (hfOut.lookAt !== null || hfOut.emote !== null) {
-        //                 contentObj.text += ". Then I ";
-        //                 if (hfOut.lookAt !== null) {
-        //                     contentObj.text += "looked at " + hfOut.lookAt;
-        //                     if (hfOut.emote !== null) {
-        //                         contentObj.text += " and ";
-        //                     }
-        //                 }
-        //                 if (hfOut.emote !== null) {
-        //                     contentObj.text = "emoted " + hfOut.emote;
-        //                 }
-        //             }
-        //
-        //             if (hfOut.actions !== null) {
-        //                 // content can only do one action
-        //                 contentObj.action = hfOut.actions[0];
-        //             }
-        //
-        //             // save response to memory
-        //             const responseMessage = {
-        //                 ...userMessage,
-        //                 userId: runtime.agentId,
-        //                 content: contentObj,
-        //             };
-        //
-        //             runtime.messageManager
-        //                 .createMemory(responseMessage)
-        //                 .then(() => {
-        //                     const messageId = stringToUuid(
-        //                         Date.now().toString()
-        //                     );
-        //                     const memory: Memory = {
-        //                         id: messageId,
-        //                         agentId: runtime.agentId,
-        //                         userId,
-        //                         roomId,
-        //                         content,
-        //                         createdAt: Date.now(),
-        //                     };
-        //
-        //                     // run evaluators (generally can be done in parallel with processActions)
-        //                     // can an evaluator modify memory? it could but currently doesn't
-        //                     runtime.evaluate(memory, state).then(() => {
-        //                         // only need to call if responseMessage.content.action is set
-        //                         if (contentObj.action) {
-        //                             // pass memory (query) to any actions to call
-        //                             runtime.processActions(
-        //                                 memory,
-        //                                 [responseMessage],
-        //                                 state,
-        //                                 async (_newMessages) => {
-        //                                     // FIXME: this is supposed override what the LLM said/decided
-        //                                     // but the promise doesn't make this possible
-        //                                     //message = newMessages;
-        //                                     return [memory];
-        //                                 }
-        //                             ); // 0.674s
-        //                         }
-        //                         resolve(true);
-        //                     });
-        //                 });
-        //         });
-        //         res.json({ response: hfOut });
-        //     }
-        // );
 
         this.app.post(
             '/:agentId/image',
